pyproject.toml
[project]
name = "infdb"
version = "0.1.0"
description = "Infrastructure database and tools for kommunale Wärmeplanung"
authors = [
  { name = "Patrick Buchenberg", email = "patrick.buchenberg@tum.de" }
]
requires-python = "==3.12.*"

dependencies = [
  "beautifulsoup4",
  "bs4", # Alias, nicht zwingend notwendig
  "chardet",
  "fastapi==0.115.7",
  "fastapi_utils==0.8.0",
  "geoalchemy2==0.17.1",
  "geopandas>=1.1.1",
  "openmeteo_requests==1.3.0",
  "pandas==2.2.3",
  "psycopg2-binary==2.9.9",
  "pydantic==2.10.5",
  "python-dotenv==1.0.1",
  "PyYAML==6.0.2",
  "requests",
  "requests_cache==1.2.1",
  "retry_requests==2.0.0",
  "SQLAlchemy==2.0.36",
  "sqlmodel==0.0.22",
  "typing_inspect==0.9.0",
  "uvicorn[standard]==0.25.0",
  "httpx==0.27.0",
]

[dependency-groups]
dev = ["flake8==7.1.1"]
docs = [
    "sphinx==7.3.7",
    "sphinx-rtd-theme==2.0.0",
    "sphinx-autodoc-typehints==2.0.0",
    "myst-parser==2.0.0"
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"
